Complete extension invocation runner ownership
