+++ crates/homeboy-core/src/extension/trace/run/runner.rs
@@ -2,8 +2,9 @@
 
 use std::path::{Path, PathBuf};
 
+use crate::extension::invoke::RunnerOutput;
 use crate::extension::resolve::ExtensionExecutionContext;
-use crate::extension::{build_scenario_runner, stderr_tail, RunnerOutput, ScenarioRunnerOptions};
+use crate::extension::{build_scenario_runner, stderr_tail, ScenarioRunnerOptions};
 use homeboy_core::component::Component;
 use homeboy_core::engine::run_dir::{self, RunDir};
 use homeboy_core::error::{Error, ErrorCode, Result};
